01_python_core/AuraFarming/requirements.txt
Pillow
opencv-python
numpy
ffpyplayer